Move `icingadb/objectpacker` to `objectpacker`
